submit for 'Apps for Food and Ag Hack 2024'' This has pending issues like lacking UI to show the flow of supply chain.
We need to create demos that prove usefulness of this application in solving the Food Recalls issue as well as incentivization of the farmers for buy-in to the network. Figuring out how the consensus can be changed from PoW to PoA network.

Move to JS stack using Solidity to write smart contracts. Using some way to simulate an actual network functioning using AWS or Ethereum's network. How to implement Consortium network, also do we actually need them in the block.